backend/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.34,<1.0
sqlalchemy>=2.0,<3.0
alembic>=1.14,<2.0
docker>=7.0,<8.0
httpx>=0.28,<1.0
pydantic>=2.0,<3.0
pydantic-settings>=2.0,<3.0
python-dotenv>=1.0,<2.0
jinja2>=3.1,<4.0
cryptography>=44.0,<45.0
bcrypt>=4.0,<6.0
PyJWT>=2.8,<3.0
